requirements.txt
selenium==4.9.0
pytest==7.4.2
python-dotenv==0.19.1
webdriver-manager==3.4.2
